pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
description="flask template"
[tool.poetry.dependencies]
flask = "==1.0.2"
python = "^3.8"
python-dotenv = "^0.15.0"
twilio = "^6.51.0"
